Corrige caminhos de importação após reorganização de pastas nos testes Cypress
